# docs/swagger.yaml
definitions:
...
  models.ServiceToken:
    properties:
      access_key:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      token_type:
        $ref: '#/definitions/models.TokenType'
      updated_at:
        type: string
    required:
    - access_key
    - name
    type: object
...
  models.TokenType:
    enum:
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - AdminToken
    - UserToken
...
  router.CreateServiceTokenRequestDto:
    properties:
      name:
        example: my-token
        type: string
    required:
    - name
    type: object
  router.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: error message
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/st/:
    get:
      description: List all service tokens (admin only).
      operationId: ListAllServiceTokens
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ServiceToken'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List all service tokens
      tags:
      - service-tokens
    post:
      consumes:
      - application/json
      description: Create a new service token. Only one admin token can exist. Only
        admin can create/delete other tokens. Admin token cannot be deleted.
      operationId: CreateServiceToken
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: Service Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateServiceTokenRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ServiceToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create service token
      tags:
      - service-tokens
  /api/v1/st/{id}:
    delete:
      description: Delete a service token by ID. Only admin can delete other tokens.
        Admin token cannot be deleted.
      operationId: DeleteServiceToken
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: Token ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Delete service token
      tags:
      - service-tokens
    get:
      description: Get a service token by ID
      operationId: GetServiceToken
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: Token ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ServiceToken'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Get service token
      tags:
      - service-tokens
  /api/v1/st/bootstrap:
    post:
      description: Create the initial admin service token. Only allowed if no admin
        token exists. Used for first-time setup.
      operationId: BootstrapAdminToken
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ServiceToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "409":
          description: Admin token already exists
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Bootstrap admin token
      tags:
      - service-tokens
  /api/v1/st/first-run:
    get:
      description: Returns whether the admin token exists. Used to determine if setup
        is required.
      operationId: FirstRunCheck
      produces:
      - application/json
      responses:
        "200":
          description: 'first_run: true if no admin token exists, false otherwise'
          schema:
            additionalProperties:
              type: boolean
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Check if admin token has already been created
      tags:
      - service-tokens
...
swagger: "2.0"